Add pessimistic DQN code
Adding the code that implements pessimism in JB's SpinningUp DQN
in TF2.

I tried to be faithful to the original pessimism implementation.
The one difference I am aware of so far is that I do not divide
by the current timestep (if I have interpreted that placeholder
correctly at all).

Plots queries over time.

Tests:
- It's not very stable, but it converges sometimes if you run it
  a few times, with a value of pessimism=10. (default).