Adding custom env & skeleton code for dist q learning
I investigated the distributed Q learning algorithm proposed.

I have sketeched out the estimators and the functions and state I
expect they will need, as a basis for discussion.

I investigated options for an environment and it looked like there
was only one way to go for this, which was to implement a custom
discrete env using the gym package. NOTE: this does not currently
implement any of the stochasticity we'll need for proving dist q
learning works as-intended. It's currently entirely deterministic.

Test plan:
 - Added unit tests testing the basic environment operation.