Account for done states, make lr more sensible
Learning rates were 1.0 which meant we were updating too heavily
on a stochastic reward function. I changed it to 0.1 default to
see what happens. We might want to consider even lower.

Also when "done" we shouldn't calculate future Q. This wasn't too
much of a problem, because we shouldn't ever be hitting "done"
states. But it's better to be correct.