Agent does not act if A_val < min r
A condition in the paper was missed. Min r is > eps_r (apart from
where it's 0), and the agent should not act if it predicts a
(normalised) Q value < eps_r. This has now been added.

Also, min reward in the stochastic reward environment is now > 0.
Before, there was a finite chance of hitting 0. reward in the
middle of the grid, which was probably incorrect.