Adding ability to train on whole history each step
We discussed a training strategy to train on the whole history each
step. This change allows us to do that.

Requires env not to reset every n steps - this is applied (though
a flag is kept in case we want the behaviour back in the future).

Also adding a way to reduce the grid size in the main script

Test plan:
 - Adding sampling-strategy option to the test_main unittests. All
   pass.