Adding update frequency argument
Pass the frequency with which to call the estimator updates as an
argument to the main.py script, useful for learning on "whole" or
"last_n_steps" sampling strategies.

Also make sampling strategies a list of choices rather than keyed
with uncomprehensible integers.

Also clean up parse args function formatting