Rename "teleport" to "event", stricter configuration
Adding empty-dict checks as the kwargs have changed names a few
times and mistakes may creep in.

We decided to go for the more general name of "event" as it doesn't
have to teleport at all. It should be more configurable than that.

For now, the dir is still called "teleporter"