Adding lr arg, final exp config, repeat track
Final final final experiment setup.

Add learning rate as an argument to the main script. 0.5 is
reasonable for convergence in the infinite env, but I don't like
changing it in the script every time!

Setting up the every_state config for the experiment. This is based
on the parameter-finding spreadsheet (can be found in the drive)

Neatening the plot to show the number of repeats