Adding newton's method of optimisation to GLNs
When convereged, Newton's method becomes viable.

Remove predict with sigma

Note - currently does it for each data point in a batch. Next, we
need to sum the hessians across the hwole batch and take the
batched inverse for the update.

Was originally going to use the equations given in the paper to
calculate grad and hessian directly, but instead use jax grad
to calculate the grad and hessian.

Tests:
  - Used in the uncertainty test, and we test fetching the weights