Fixing environment bugs
Clip had been implemented incorrectly, so that the cart could still
fail (this was not intentional).

Also, if it fails on a knocking step, there was no way to determine
how the agent failed. This has been fixed by stopping knocking
happening if the episode has just ended that step.

Tests:
 - None